refactor: typescriptify git branch file resolver
