+++ projects/openapi-io/src/parser/types.ts
@@ -1,0 +1,5 @@
+export type ExternalRefHandler = {
+  order: number;
+  canRead: (file: { url: string }) => boolean;
+  read: (file: { url: string }) => Promise<string>;
+};
